move internal urls to their own file for exporting, clarity
