+++ frontend/src/pages/DataCatalog/AgeAdjustmentTab.tsx
@@ -6,7 +6,7 @@ import {
   COVID_DEATHS_US_SETTING,
   COVID_HOSP_US_SETTING,
   EXPLORE_DATA_PAGE_LINK,
-} from "../../utils/urlutils";
+} from "../../utils/internalRoutes";
 import { Link } from "react-router-dom";
 import { Button } from "@material-ui/core";
