+++ frontend/src/pages/TermsOfUsePage/TermsOfUsePage.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import Grid from "@material-ui/core/Grid";
 import styles from "./TermsOfUsePage.module.scss";
 import { Typography } from "@material-ui/core";
-import { HET_URL } from "../../utils/urlutils";
+import { HET_URL } from "../../utils/internalRoutes";
 import { Helmet } from "react-helmet-async";
 
 function TermsOfUsePage() {
