+++ frontend/src/pages/WhatIsHealthEquity/News/NewsPreviewCard.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import { NEWS_TAB_LINK, ReactRouterLinkButton } from "../../../utils/urlutils";
+import { NEWS_TAB_LINK } from "../../../utils/internalRoutes";
 import styles from "./News.module.scss";
 import AppbarLogo from "../../../assets/AppbarLogo.png";
-import { getHtml } from "../../../utils/urlutils";
+import { getHtml, ReactRouterLinkButton } from "../../../utils/urlutils";
 import { Article } from "../NewsTab";
 import { Box, Grid } from "@material-ui/core";
 import LazyLoad from "react-lazyload";
