+++ frontend/src/pages/WhatIsHealthEquity/NewsTab.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styles from "./WhatIsHealthEquityPage.module.scss";
 import Grid from "@material-ui/core/Grid";
-import { NEWS_TAB_LINK } from "../../utils/urlutils";
+import { NEWS_TAB_LINK } from "../../utils/internalRoutes";
 import { Route, Switch } from "react-router-dom";
 import AllPosts from "./News/AllPosts";
 import SinglePost from "./News/SinglePost";
