+++ frontend/src/pages/ui/FaqSection.tsx
@@ -5,7 +5,9 @@ import Typography from "@material-ui/core/Typography";
 import { Accordion, AccordionSummary } from "@material-ui/core";
 import AccordionDetails from "@material-ui/core/AccordionDetails";
 import ExpandMoreIcon from "@material-ui/icons/ExpandMore";
-import { FAQ_TAB_LINK, ReactRouterLinkButton } from "../../utils/urlutils";
+import { FAQ_TAB_LINK } from "../../utils/internalRoutes";
+import { ReactRouterLinkButton } from "../../utils/urlutils";
+
 import { selectFaqs } from "../WhatIsHealthEquity/FaqTab";
 import { getHtml } from "../../utils/urlutils";
