+++ frontend/src/pages/ui/SignupSection.tsx
@@ -1,7 +1,9 @@
 import { Button, Grid, TextField } from "@material-ui/core";
 import React from "react";
 import { urlMap } from "../../utils/externalUrls";
-import { CONTACT_TAB_LINK, LinkWithStickyParams } from "../../utils/urlutils";
+import { LinkWithStickyParams } from "../../utils/urlutils";
+import { CONTACT_TAB_LINK } from "../../utils/internalRoutes";
+
 import styles from "./SignupSection.module.scss";
 
 export default function SignupSection() {
