+++ frontend/src/utils/internalRoutes.ts
@@ -1,0 +1,39 @@
+// PAGE URLS
+export const HET_URL = "https://healthequitytracker.org";
+
+export const EXPLORE_DATA_PAGE_LINK = "/exploredata";
+export const DATA_CATALOG_PAGE_LINK = "/datacatalog";
+export const ABOUT_US_PAGE_LINK = "/aboutus";
+export const WHAT_IS_HEALTH_EQUITY_PAGE_LINK = "/whatishealthequity";
+export const TERMS_OF_USE_PAGE_LINK = "/termsofuse";
+
+// TAB URLS
+export const FAQ_TAB_LINK = "/faqs";
+export const RESOURCES_TAB_LINK = "/resources";
+export const METHODOLOGY_TAB_LINK = "/methodology";
+export const AGE_ADJUSTMENT_TAB_LINK = "/ageadjustment";
+export const DATA_TAB_LINK = "/datacatalog";
+export const CONTACT_TAB_LINK = "/contact";
+export const OURTEAM_TAB_LINK = "/ourteam";
+export const NEWS_TAB_LINK = "/news";
+
+// TRACKER SETTINGS
+export const COVID_DEATHS_US_SETTING =
+  "?mls=1.covid-3.00&dt1=covid_deaths&demo=race_and_ethnicity";
+export const COVID_HOSP_US_SETTING =
+  "?mls=1.covid-3.00&dt1=covid_hospitalizations&demo=race_and_ethnicity";
+export const COVID_HOSP_NY_COUNTY_SETTING =
+  "?dt1=covid_hospitalizations&mls=1.covid-3.36061";
+export const COVID_VAX_US_SETTING = "?mls=1.covid_vaccinations-3.00";
+export const COPD_US_SETTING = "?mls=1.copd-3.00";
+export const DIABETES_US_SETTING = "?mls=1.diabetes-3.00";
+export const UNINSURANCE_US_SETTING = "?mls=1.health_insurance-3.00";
+export const POVERTY_US_SETTING = "?mls=1.poverty-3.00";
+export const OPIOID_US_SETTING =
+  "?dt1=non_medical_drug_use&mls=1.substance-3.00";
+
+// SECTION IDS
+export const WHAT_DATA_ARE_MISSING_ID = "missingDataInfo";
+export const EXPLORE_DATA_PAGE_WHAT_DATA_ARE_MISSING_LINK =
+  EXPLORE_DATA_PAGE_LINK + "#" + WHAT_DATA_ARE_MISSING_ID;
+export const WIHE_JOIN_THE_EFFORT_SECTION_ID = "join";
